Aligned impl with JIT 1.2 specs.

+++ jit-persistence/src/main/java/org/dcsa/jit/persistence/entity/OperationsEvent.java
@@ -79,7 +79,7 @@ public class OperationsEvent {
   private FacilityTypeCodeOPR facilityTypeCode;
 
   @Column(name = "vessel_draft")
-  private Integer vesselDraft;
+  private Float vesselDraft;
 
   @Enumerated(EnumType.STRING)
   @Column(name = "vessel_draft_unit")
@@ -87,5 +87,5 @@ public class OperationsEvent {
   private DimensionUnit vesselDraftUnit;
 
   @Column(name = "miles_remaining_to_destination")
-  private Integer milesRemainingToDestination;
+  private Float milesRemainingToDestination;
 }
